Fixed header bug
