+++ components/Header.vue
@@ -57,7 +57,21 @@
             <div class="hidden md:flex items-center justify-end flex-1 md:w-0">
               <!-- TODO: Site search-->
             </div>
-            <ThemeSwitcher :selected="selected" />
+            <div
+              class="text-gray-500 dark:text-dark-text hover:bg-gray-200 dark:hover:bg-gray-700 focus:outline-none focus:ring-4 focus:ring-gray-200 dark:focus:ring-gray-700 rounded-lg text-sm p-2.5">
+              <div class="relative flex items-center ">
+                <SunIcon :class="selected == 'Light' ? 'h-6 w-6' : 'hidden'" />
+                <MoonIcon :class="selected == 'Dark' ? 'h-6 w-6' : 'hidden'" />
+                <DesktopComputerIcon :class="selected == 'System' ? 'h-6 w-6' : 'hidden'" />
+                <select name="theme" id="theme"
+                  class="absolute appearance-none pr-3 w-full h-fit opacity-0 dark:bg-slate-600 dark:text-white"
+                  @change="switchSelect($event)">
+                  <option :selected="selected == 'Light'">Light</option>
+                  <option :selected="selected == 'Dark'">Dark</option>
+                  <option :selected="selected == 'System'">System</option>
+                </select>
+              </div>
+            </div>
           </div>
         </div>
       </div>
@@ -99,7 +113,22 @@
                 </div>
               </div>
               <div class="flex items-center justify-between py-6 px-3">
-                <MobileThemeSwitcher :selected="selected" />
+                <label class="text-gray-900 text-base font-medium dark:text-mitre-silver">Switch Theme</label>
+                <div
+                  class="relative flex items-center ring-1 ring-slate-900/10 rounded-lg shadow-sm p-2 text-slate-700 font-semibold dark:bg-slate-600 dark:ring-0 dark:highlight-white/5 dark:text-slate-200">
+                  <SunIcon :class="selected == 'Light' ? 'h-6 w-6' : 'hidden'" />
+                  <MoonIcon :class="selected == 'Dark' ? 'h-6 w-6' : 'hidden'" />
+                  <DesktopComputerIcon :class="selected == 'System' ? 'h-6 w-6' : 'hidden'" />
+                  <div class="ml-2">{{ selected }}</div>
+                  <ChevronDownIcon class="w-5 h-6  text-slate-400" />
+                  <select id="theme"
+                    class="absolute appearance-none inset-0 w-full h-full opacity-0 dark:bg-slate-600 dark:text-white"
+                    @change="switchSelect($event)">
+                    <option :selected="selected == 'Light'">Light</option>
+                    <option :selected="selected == 'Dark'">Dark</option>
+                    <option :selected="selected == 'System'">System</option>
+                  </select>
+                </div>
               </div>
             </div>
           </PopoverPanel>
@@ -125,7 +154,10 @@ import {
   XIcon,
   ClipboardCheckIcon,
   DotsVerticalIcon,
-  UserGroupIcon
+  UserGroupIcon,
+  DesktopComputerIcon,
+  MoonIcon,
+  SunIcon,
 } from '@heroicons/vue/outline';
 import { AcademicCapIcon, BookOpenIcon, RssIcon } from '@heroicons/vue/solid';
 
@@ -281,6 +313,33 @@ const resources = ref([
 
 
 ////  Methods  ////
+const switchSelect = (event) => {
+  if (event.target.value === 'Light') {
+    disableDarkMode()
+    localStorage.setItem('user-theme', 'light');
+    selected.value = "Light"
+    setThemeState(selected.value)
+  }
+  else if (event.target.value == 'Dark') {
+    enableDarkMode()
+    localStorage.setItem('user-theme', 'dark');
+    selected.value = "Dark"
+    setThemeState(selected.value)
+  }
+  else if (event.target.value == 'System') {
+    const userIsInDarkModeOS = window.matchMedia('(prefers-color-scheme: dark)').matches
+    if (userIsInDarkModeOS) {
+      enableDarkMode()
+      localStorage.setItem('user-theme', 'dark');
+      selected.value = "System";
+      setThemeState(selected.value)
+    }
+    else {
+      disableDarkMode()
+      selected.value = "System";
+    }
+  }
+}
 const setThemeState = (mode) => {
   localStorage.setItem('theme', mode)
 }
